requirements.txt
paho-mqtt==1.4.0
Flask==1.0.2
